all API calls implemented with Vault in dev mode
