+++ cli/example.json
@@ -1,0 +1,7 @@
+{
+    "name": "user",
+    "metadata": {
+        "dataset": 4,
+        "dataset2": 3
+    }
+}
