--- docs/cli/docker-compose.yml
@@ -1,15 +1,0 @@
-version: '2'
-services:
-  vault:
-    image: vault
-    container_name: vault
-    ports:
-      - "8200:8200"
-    restart: always
-    volumes:
-      - ./volumes/logs:/vault/logs
-      - ./volumes/file:/vault/file
-      - ./volumes/config:/vault/config
-    cap_add:
-      - IPC_LOCK
-    entrypoint: vault server -config=/vault/config/vault.json
