--- docs/cli/example.json
@@ -1,7 +1,0 @@
-{
-    "name": "user",
-    "metadata": {
-        "dataset": 4,
-        "dataset2": 3
-    }
-}
