--- docs/cli/volumes/config/vault.json
@@ -1,14 +1,0 @@
-{
-  "backend": {
-    "file": {
-      "path": "/vault/file"
-    }
-  },
-  "listener": {
-    "tcp":{
-      "address": "0.0.0.0:8200",
-      "tls_disable": 1
-    }
-  },
-  "ui": true
-}
